Clarification
Rephrasing 
"Requires occasional support from peers depending on the task, technology and prior experience." to
"Only occasionally requires support from peers depending on the task, technology and prior experience" to make it clearer that the goal is that the individual needs _less_ support. Someone who needs no support is meeting the expectation.